requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.1
psycopg2==2.9.4
PyJWT==2.5.0
pyotp==2.7.0
python-dotenv==0.21.0
pytz==2022.4
sqlparse==0.4.3
